# Create empty bloom filter only once