For some reason, `~BloomFilter(BLOOM_FILTER_SIZE);` is quite expensive, resulting in a `> 1 ms` startup time for a join. For as long as we do not have dynamic bloom filter sizes, this is a hotfix:

**System**
<details>
<summary>nemea - click to expand</summary>

| property | value |
| -- | -- |
| Hostname | nemea |
| CPU | Intel(R) Xeon(R) Platinum 8180 CPU @ 2.50GHz |
| Memory | 1.2TB |
| numactl | nodebind: 0  |
| numactl | membind: 0  |
</details>

**Commit Info and Build Time**
| commit | date | message | build time |
| -- | -- | -- | -- |
| 88b690fc4 | 07.10.2020 07:35 | benchmark_all: Include TPC-H SF 10 | real 265.11 user 2399.13 sys 132.74|
| 792e6f29a | 07.10.2020 07:35 | Reuse empty bloom filter | real 266.38 user 2385.16 sys 135.59|


**hyriseBenchmarkTPCH - single-threaded**
<details>
<summary>
Sum of avg. item runtimes: +1%
 ||
Geometric mean of throughput changes: +0%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise3/build-release/benchmark_all_results/hyriseBenchmarkTPCH_88b690fc4f1881f9a0dfea42caf2946881517848_st.json | /home/Markus.Dreseler/hyrise3/build-release/benchmark_all_results/hyriseBenchmarkTPCH_792e6f29a93aa55cc6f22ecf05c5061d7e3d64d0_st.json |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 88b690fc4f1881f9a0dfea42caf2946881517848-dirty                                                                                         | 792e6f29a93aa55cc6f22ecf05c5061d7e3d64d0-dirty                                                                                         |
 |  benchmark_mode          | Ordered                                                                                                                                | Ordered                                                                                                                                |
 |  build_type              | release                                                                                                                                | release                                                                                                                                |
 |  chunk_size              | 65535                                                                                                                                  | 65535                                                                                                                                  |
 |  clients                 | 1                                                                                                                                      | 1                                                                                                                                      |
 |  compiler                | gcc 9.2                                                                                                                                | gcc 9.2                                                                                                                                |
 |  cores                   | 0                                                                                                                                      | 0                                                                                                                                      |
 |  date                    | 2020-10-07 07:40:16                                                                                                                    | 2020-10-07 14:37:06                                                                                                                    |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                | {'default': {'encoding': 'Dictionary'}}                                                                                                |
 |  indexes                 | False                                                                                                                                  | False                                                                                                                                  |
 |  max_duration            | 60000000000                                                                                                                            | 60000000000                                                                                                                            |
 |  max_runs                | -1                                                                                                                                     | -1                                                                                                                                     |
 |  scale_factor            | 1.0                                                                                                                                    | 1.0                                                                                                                                    |
 |  time_unit               | ns                                                                                                                                     | ns                                                                                                                                     |
 |  use_prepared_statements | False                                                                                                                                  | False                                                                                                                                  |
 |  using_scheduler         | False                                                                                                                                  | False                                                                                                                                  |
 |  verify                  | False                                                                                                                                  | False                                                                                                                                  |
 |  warmup_duration         | 0                                                                                                                                      | 0                                                                                                                                      |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||    815.5 |   823.1 |   +1%  ||     1.23 |     1.21 |   -1%  |  0.1388 |
+| TPC-H 02 ||      5.1 |     4.9 |   -5%  ||   194.69 |   204.39 |   +5%  |  0.0000 |
 | TPC-H 03 ||     80.5 |    81.0 |   +1%  ||    12.43 |    12.35 |   -1%  |  0.0000 |
 | TPC-H 04 ||     68.1 |    68.7 |   +1%  ||    14.68 |    14.56 |   -1%  |  0.0000 |
-| TPC-H 05 ||    217.1 |   229.8 |   +6%  ||     4.61 |     4.35 |   -6%  |  0.0000 |
+| TPC-H 06 ||      3.4 |     3.2 |   -5%  ||   297.91 |   313.21 |   +5%  |  0.0000 |
 | TPC-H 07 ||     58.0 |    58.2 |   +0%  ||    17.24 |    17.20 |   -0%  |  0.7748 |
 | TPC-H 08 ||     68.6 |    68.5 |   -0%  ||    14.57 |    14.60 |   +0%  |  0.6851 |
 | TPC-H 09 ||    481.7 |   482.1 |   +0%  ||     2.08 |     2.07 |   -0%  |  0.7908 |
 | TPC-H 10 ||    167.3 |   167.9 |   +0%  ||     5.98 |     5.96 |   -0%  |  0.5086 |
 | TPC-H 11 ||      9.1 |     9.1 |   +0%  ||   110.32 |   110.24 |   -0%  |  0.5013 |
 | TPC-H 12 ||     44.1 |    43.6 |   -1%  ||    22.66 |    22.95 |   +1%  |  0.0000 |
 | TPC-H 13 ||    354.9 |   365.0 |   +3%  ||     2.82 |     2.74 |   -3%  |  0.0000 |
 | TPC-H 14 ||     20.8 |    21.1 |   +1%  ||    48.02 |    47.40 |   -1%  |  0.0000 |
 | TPC-H 15 ||      7.8 |     7.7 |   -1%  ||   127.63 |   129.16 |   +1%  |  0.0000 |
 | TPC-H 16 ||     94.3 |    91.0 |   -4%  ||    10.60 |    10.99 |   +4%  |  0.0000 |
 | TPC-H 17 ||     18.0 |    17.9 |   -1%  ||    55.43 |    55.86 |   +1%  |  0.0001 |
 | TPC-H 18 ||    688.4 |   697.8 |   +1%  ||     1.45 |     1.43 |   -1%  |  0.1728 |
 | TPC-H 19 ||     29.7 |    29.3 |   -1%  ||    33.70 |    34.15 |   +1%  |  0.0000 |
 | TPC-H 20 ||     16.2 |    16.1 |   -0%  ||    61.83 |    61.97 |   +0%  |  0.4158 |
 | TPC-H 21 ||    420.6 |   421.8 |   +0%  ||     2.38 |     2.37 |   -0%  |  0.5384 |
 | TPC-H 22 ||     49.4 |    49.5 |   +0%  ||    20.22 |    20.20 |   -0%  |  0.0122 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||   3718.6 |  3757.3 |   +1%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   +0%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```
</details>


**hyriseBenchmarkTPCH - single-threaded, SF 0.01**
<details>
<summary>
Sum of avg. item runtimes: -2%
 ||
Geometric mean of throughput changes: +5%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise3/build-release/benchmark_all_results/hyriseBenchmarkTPCH_88b690fc4f1881f9a0dfea42caf2946881517848_st_s01.json | /home/Markus.Dreseler/hyrise3/build-release/benchmark_all_results/hyriseBenchmarkTPCH_792e6f29a93aa55cc6f22ecf05c5061d7e3d64d0_st_s01.json |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 88b690fc4f1881f9a0dfea42caf2946881517848-dirty                                                                                             | 792e6f29a93aa55cc6f22ecf05c5061d7e3d64d0-dirty                                                                                             |
 |  benchmark_mode          | Ordered                                                                                                                                    | Ordered                                                                                                                                    |
 |  build_type              | release                                                                                                                                    | release                                                                                                                                    |
 |  chunk_size              | 65535                                                                                                                                      | 65535                                                                                                                                      |
 |  clients                 | 1                                                                                                                                          | 1                                                                                                                                          |
 |  compiler                | gcc 9.2                                                                                                                                    | gcc 9.2                                                                                                                                    |
 |  cores                   | 0                                                                                                                                          | 0                                                                                                                                          |
 |  date                    | 2020-10-07 08:02:34                                                                                                                        | 2020-10-07 14:59:24                                                                                                                        |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                    | {'default': {'encoding': 'Dictionary'}}                                                                                                    |
 |  indexes                 | False                                                                                                                                      | False                                                                                                                                      |
 |  max_duration            | 60000000000                                                                                                                                | 60000000000                                                                                                                                |
 |  max_runs                | -1                                                                                                                                         | -1                                                                                                                                         |
 |  scale_factor            | 0.009999999776482582                                                                                                                       | 0.009999999776482582                                                                                                                       |
 |  time_unit               | ns                                                                                                                                         | ns                                                                                                                                         |
 |  use_prepared_statements | False                                                                                                                                      | False                                                                                                                                      |
 |  using_scheduler         | False                                                                                                                                      | False                                                                                                                                      |
 |  verify                  | False                                                                                                                                      | False                                                                                                                                      |
 |  warmup_duration         | 0                                                                                                                                          | 0                                                                                                                                          |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||      7.3 |     7.3 |   -0%  ||   137.54 |   137.60 |   +0%  |  0.4183 |
+| TPC-H 02 ||      0.7 |     0.5 |  -17%  ||  1527.95 |  1849.06 |  +21%  |  0.0000 |
 | TPC-H 03 ||      0.9 |     0.9 |   -2%  ||  1098.37 |  1121.84 |   +2%  |  0.0000 |
+| TPC-H 04 ||      0.5 |     0.5 |   -5%  ||  1831.17 |  1918.32 |   +5%  |  0.0000 |
+| TPC-H 05 ||      1.3 |     1.2 |   -6%  ||   760.05 |   805.82 |   +6%  |  0.0000 |
 | TPC-H 06 ||      0.1 |     0.1 |   -1%  ||  9327.17 |  9403.38 |   +1%  |  0.0000 |
+| TPC-H 07 ||      1.3 |     1.2 |   -8%  ||   744.20 |   810.05 |   +9%  |  0.0000 |
 | TPC-H 08 ||     14.8 |    14.5 |   -2%  ||    67.49 |    68.78 |   +2%  |  0.0159 |
 | TPC-H 09 ||      2.4 |     2.4 |   -2%  ||   413.18 |   420.34 |   +2%  |  0.0000 |
+| TPC-H 10 ||      1.0 |     0.9 |   -5%  ||  1042.21 |  1093.16 |   +5%  |  0.0000 |
+| TPC-H 11 ||      0.3 |     0.3 |  -14%  ||  2894.60 |  3372.29 |  +17%  |  0.0000 |
 | TPC-H 12 ||      0.7 |     0.6 |   -3%  ||  1519.77 |  1568.84 |   +3%  |  0.0000 |
 | TPC-H 13 ||      2.2 |     2.3 |   +1%  ||   446.70 |   442.89 |   -1%  |  0.0000 |
+| TPC-H 14 ||      0.4 |     0.4 |   -6%  ||  2625.98 |  2806.57 |   +7%  |  0.0000 |
 | TPC-H 15 ||      1.2 |     1.1 |   -3%  ||   865.37 |   892.13 |   +3%  |  0.0000 |
 | TPC-H 16 ||      2.2 |     2.1 |   -3%  ||   453.33 |   465.37 |   +3%  |  0.0000 |
+| TPC-H 17 ||      0.3 |     0.3 |  -11%  ||  3040.84 |  3433.56 |  +13%  |  0.0000 |
 | TPC-H 18 ||      2.7 |     2.7 |   -2%  ||   363.97 |   370.04 |   +2%  |  0.0000 |
 | TPC-H 19 ||      5.5 |     5.5 |   -1%  ||   181.07 |   182.59 |   +1%  |  0.0000 |
 | TPC-H 20 ||      2.4 |     2.4 |   -2%  ||   414.93 |   424.94 |   +2%  |  0.0000 |
 | TPC-H 21 ||      2.2 |     2.1 |   -3%  ||   455.55 |   470.85 |   +3%  |  0.0000 |
 | TPC-H 22 ||      1.3 |     1.3 |   -0%  ||   768.69 |   772.04 |   +0%  |  0.0000 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
…